requirements.txt
asgiref==3.3.4
cx-Oracle==8.1.0
Django==3.2
django-admin-interface==0.16.3
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-filter==2.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
Pillow==8.2.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
